unsigned / signed fix .. attempt #2

+++ src/doom/r_plane.cpp
@@ -472,7 +472,7 @@ void R_DrawPlanes(void)
                 dc_yl = pl->top[x];
                 dc_yh = pl->bottom[x];
 
-                if (static_cast<unsigned>(dc_yl) <= dc_yh) // [crispy] 32-bit integer math
+                if (static_cast<unsigned>(dc_yl) <= static_cast<unsigned>(dc_yh)) // [crispy] 32-bit integer math
                 {
                     angle     = ((an + xtoviewangle[x]) ^ flip) >> ANGLETOSKYSHIFT;
                     dc_x      = x;
